dbus-broker: drop live-system conversion
With F33 released, we now have 4 Fedora releases (F30-F33) that contain
a live-upgrade of `dbus-daemon` to `dbus-broker`. This allowed us to
convert existing machines out there to the new bus implementation.

However, the live-conversion is fragile, since it requires relinking
systemd services at runtime, tricking it into believing they did not
change. This is not guaranteed to work forever, nor do we want to rely
on it. Hence, this drops this code-path.

This effectively means that if you update pre-F29 machines to post-F34,
you will continue using `dbus-daemon`. This is completely fine and a
consistent *and* supported system setup. However, you no longer get the
benefits of dbus-broker automatically. Instead, you must explicitly
switch from dbus-daemon to dbus-broker. The easiest way to do that is
to install dbus-broker and uninstall dbus-daemon. This will
automatically convert the services to the new daemon.

Lastly, any fresh Fedora install will now always pick up dbus-broker,
since the `fedora-release` presets are merged and correctly configure
dbus-broker. Hence, this change only affects old installs.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>